merge_pcap: cleaning fix
The tmp.pcap moving should not be done because it is tshark that
generates the output file. Furthermore, the path tmp.pcap was wrong.
tmp.pcap is inside the "WORK_DIR" directory and not in the current
directory.

Also, the `rm $WORK_DIR/*` was useless because the directory is removed
afterward.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>