merge_pcap: support for abritrary pcap duration
In the previous code the pcap a and b files were supposed to have a
duration of 1 second. Also, there was no delay between the two files.

This commit adds the possibility to specify the frequency of the
current to add the appropriate delay between the two files. In addition,
it checks the duration of the pcap files to calculate the offset instead
of hardcoding it to 1 second.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>